Login and Logout and setupLocator
