+++ lib/main.dart
@@ -1,3 +1,4 @@
+import 'package:demo_app/provider/getit.dart';
 import 'package:demo_app/route_generator.dart';
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
@@ -8,6 +9,7 @@ Future main() async {
   await Firebase.initializeApp(
     options: DefaultFirebaseOptions.currentPlatform,
   );
+  setupLocator();
   runApp(
     const MaterialApp(
       debugShowCheckedModeBanner: false,
