+++ lib/view/register_viewmodel.dart
@@ -1,3 +1,4 @@
+import 'package:demo_app/provider/base_model.dart';
 import 'package:flutter/material.dart';
 
-class RegisterViewModel {}
+class RegisterViewModel extends BaseModel {}
