+++ pubspec.yaml
@@ -37,6 +37,7 @@ dependencies:
   firebase_core: ^1.12.0
   firebase_auth: ^3.3.8
   get_it: ^7.2.0
+  provider: ^6.0.3
 
 dev_dependencies:
   flutter_test:
